🗺️ Atlas: Refactor CurrentStorage to reduce Blob anti-pattern
- Moved `src/storage/current.rs` to `src/storage/current/mod.rs`
- Extracted `iterators.rs` for edge iterators
- Extracted `stats.rs` for `CurrentStats` and `FilterStats`
- Extracted `vector.rs` for vector index helpers
- Extracted `tests.rs` for unit tests
- Updated imports and visibility
